Implement use-case approach with Screaming architecture. This will allow to comply with OCP
